+++ src/Buildalyzer/AnalyzerManagerOptions.cs
@@ -1,6 +1,8 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using Buildalyzer.Construction;
 using Buildalyzer.Logging;
+using Microsoft.Build.Construction;
 using Microsoft.Build.Framework;
 using Microsoft.Extensions.Logging;
 
@@ -10,6 +12,12 @@ namespace Buildalyzer
     {
         public ILoggerFactory LoggerFactory { get; set; }
         public IProjectTransformer ProjectTransformer { get; set; }
+
+        /// <summary>
+        /// A filter that indicates whether a give project should be loaded.
+        /// Return <c>true</c> to load the project, <c>false</c> to filter it out.
+        /// </summary>
+        public Func<ProjectInSolution, bool> ProjectFilter { get; set; }
 
         public TextWriter LogWriter
         {